moving proactiveEventHandler, AudoPlayerEventHandler, SessionEndedRequestHandler, ErrorHandler to helperIntents file
